Fixed bug: immediately adding chat to screen
In the addMessageToStore function, the function would directly modify the state when appending messages to the convo array. Therefore, the state would not recognize the change because the previous state and next state were the same. By creating a copy of the convo in the map function, the state will update properly and re-render the message components when the props are changed. Changed append() to unshift() to allow for correct message order when sending a message. Made similar changes to addNewConvoToStore function to fix bug when sending message to a new recipient would not display immediately.